+++ src/cli.rs
@@ -298,7 +298,7 @@ pub struct Opts {
     #[arg(
         long,
         short = 'E',
-        value_name = "pattern",
+        value_name = "glob",
         help = "Exclude entries that match the given glob pattern",
         long_help
     )]
@@ -526,6 +526,10 @@ pub struct Opts {
         help = "Add hyperlinks to output paths"
     )]
     pub hyperlink: HyperlinkWhen,
+
+    /// Ignore directories containing the named entry.
+    #[arg(long, value_name = "name")]
+    pub ignore_contain: Vec<String>,
 
     /// Set number of threads to use for searching & executing (default: number
     /// of available CPU cores)
@@ -867,7 +871,8 @@ impl clap::Args for Exec {
                     "Execute a command for each search result in parallel (use --threads=1 for sequential command execution). \
                      There is no guarantee of the order commands are executed in, and the order should not be depended upon. \
                      All positional arguments following --exec are considered to be arguments to the command - not to fd. \
-                     It is therefore recommended to place the '-x'/'--exec' option last.\n\
+                     It is therefore recommended to place the '-x'/'--exec' option last. \
+                     Use '\\;' to terminate the command template if you need to continue passing fd arguments afterwards.\n\
                      The following placeholders are substituted before the command is executed:\n  \
                        '{}':   path (of the current search result)\n  \
                        '{/}':  basename\n  \
@@ -882,6 +887,8 @@ impl clap::Args for Exec {
                            fd -e zip -x unzip\n\n  \
                        - find *.h and *.cpp files and run \"clang-format -i ..\" for each of them:\n\n      \
                            fd -e h -e cpp -x clang-format -i\n\n  \
+                       - search within `src/` and echo each match (place `-x` last):\n\n      \
+                           fd . src -x echo\n\n  \
                        - Convert all *.jpg files to *.png files:\n\n      \
                            fd -e jpg -x convert {} {.}.png\
                     ",